Extracted SQL builder from execution on collection to enable unit testing SQL generation in the future
